Modify server request factory
The PSR-7 specification provides no way to update the server parameters
after a server request is instantiated, so the server parameters must be
provided directly. Because the request method and URI are typically
included in the server parameters, make these optional for overloading
the server values.

Having a server factory that creates from globals is pointless as an
interface, since this functionality is universal and can be applied to
any implementation that uses the server request factory.

Refs #18
Refs #20
Refs #21