Remove inheritance from all interfaces
Based on discussions within the working group, this patch removes
inheritance from all proposed interfaces.

We may create union types in the future (interfaces that extend multiple
interfaces in order to provide a single type with factory methods for
all collaborators), but for now, consensus is that the base interfaces
should not do so.